pickle_2

+++ src/read_data.py
@@ -40,24 +40,33 @@ datapat = re.compile(r'.*(?P<datetime>[0-9]{4}[\\-][0-9]{2}[\\-][0-9]{2}.*[0-9]{
 
 puncpat = re.compile(r'[ -.\\:]')
 
-def read_data(indexlist):
+def read_to_database(indexlist):
     for i in indexlist:
-        res = config.data_format.copy()
-        with open(config.data_file_list[i], 'r', encoding = 'utf-8') as fp:
-            print('---', config.data_file_list[i], '---')
-            for line in fp.readlines():
-                dat = re.match(datapat, line).groupdict()
-                if not dat:
-                    print(line, 'can not be parsed')
-                    continue
-                if not dat['id'] in config.contract_list:
-                    print(dat['id'], 'is not in constract list')
-                    continue
-                dat['datetime'] = datetime.datetime(*[int(x) for x in re.split(puncpat, dat['datetime']+'000')])
-                res[dat['id']].append(dat)
-        with open(__database_dir__ + str(i), 'wb') as dbf:
-            pickle.dump(res, dbf)
-            dbf.close()
+        if os.path.exists(config.data_file_list[i]) == false:
+            res = config.data_format.copy()
+            with open(config.data_file_list[i], 'r', encoding='utf-8') as fp:
+                print('---', config.data_file_list[i], '---')
+                for line in fp.readlines():
+                    dat = re.match(datapat, line).groupdict()
+                    if not dat:
+                        print(line, 'can not be parsed')
+                        continue
+                    if not dat['id'] in config.contract_list:
+                        print(dat['id'], 'is not in constract list')
+                        continue
+                    dat['datetime'] = datetime.datetime(*[int(x) for x in re.split(puncpat, dat['datetime'] + '000')])
+                    res[dat['id']].append(dat)
+            with open(__database_dir__ + str(i), 'wb') as dbf:
+                pickle.dump(res, dbf)
+                dbf.close()
+def read_data(indexlist):
+    res = config.data_format.copy()
+    tmp = config.data_format.copy()
+    for i in indexlist:
+        with open(__database_dir__ + str(i), 'rb') as dbf:
+            tmp = pickle.load(dbf)
+        for key in res.
+
 
 if __name__ == '__main__':
     # for test
